refactor: extraction massive composants + corrections audit technique
Phase 1 — Critiques:
- SplashScreen: 501 → 72 lignes (suppression 100 particules, shockwaves)
- registerSW.ts: console.error gaté en DEV
- 20+ console.log/warn/error gatés derrière import.meta.env.DEV
- Validation whitelist domaines iframe (videoEmbed + embed)

Phase 2 — Composants partagés:
- 8 widgets sidebar extraits (TableOfContents, Share, Video, Tags,
  Newsletter, Popular, Related, CTA) → src/components/sidebar/
- PortraitSidebar + PortraitNavigationFooter → src/components/portrait/
- RelatedArticles full-width factorisé (article + video)
- Navbar 1267 → 173 lignes + 11 sous-composants → src/components/navbar/
- Hooks partagés useShare + useScrollProgress → src/hooks/

Phase 2 — Design system:
- 9 boutons inline convertis vers <Button /> / <IconButton />
  (HeroSection, UniversSection, Footer)

Résultat:
- ArticlePage: 1181 → 783 (-34%)
- VideoPage: 1148 → 737 (-36%)
- PortraitDetailPage: 1040 → 710 (-32%)
- Navbar: 1267 → 173 (-86%)
- SplashScreen: 501 → 72 (-86%)
- Net: -2566 lignes, +26 composants/hooks partagés
- Build OK, 23 tests passent
