+++ src/components/navbar/index.ts
@@ -4,3 +4,15 @@
 export * from './types';
 export * from './SocialIcons';
 export * from './navData';
+
+export { default as DesktopTopBar } from './DesktopTopBar';
+export { default as DesktopNav } from './DesktopNav';
+export { default as MegaMenuArticles } from './MegaMenuArticles';
+export { default as MegaMenuVideos } from './MegaMenuVideos';
+export { default as MegaMenuHistoires } from './MegaMenuHistoires';
+export { default as MegaMenuSeries } from './MegaMenuSeries';
+export { default as MegaMenuBoutique } from './MegaMenuBoutique';
+export { default as MegaMenuRecos } from './MegaMenuRecos';
+export { default as MobileHeader } from './MobileHeader';
+export { default as MobileSearchBar } from './MobileSearchBar';
+export { default as MobileMenu } from './MobileMenu';
