+++ src/components/portrait/index.ts
@@ -1,0 +1,4 @@
+// src/components/portrait/index.ts
+export * from './types';
+export { default as PortraitSidebar } from './PortraitSidebar';
+export { default as PortraitNavigationFooter } from './PortraitNavigationFooter';
