+++ src/components/sidebar/index.ts
@@ -1,0 +1,9 @@
+// src/components/sidebar/index.ts
+export { default as SidebarTableOfContents } from './SidebarTableOfContents';
+export { default as SidebarShare } from './SidebarShare';
+export { default as SidebarVideo } from './SidebarVideo';
+export { default as SidebarTags } from './SidebarTags';
+export { default as SidebarNewsletter } from './SidebarNewsletter';
+export { default as SidebarPopular } from './SidebarPopular';
+export { default as SidebarRelated } from './SidebarRelated';
+export { default as SidebarCTA } from './SidebarCTA';
